requirements.txt
Django==6.0.3
dj-database-url==3.1.2
python-dotenv==1.0.1
psycopg2-binary==2.9.12
whitenoise==6.12.0
gunicorn==25.3.0
pillow==11.1.0
pandas==2.2.3
reportlab==4.4.10
tensorflow==2.21.0
